Change timeouts to double type everywhere

+++ src/SharedMemoryUser.hpp
@@ -65,10 +65,10 @@ namespace geopm
             virtual std::unique_ptr<SharedMemoryScopedLock> get_scoped_lock(void) = 0;
             /// @brief Returns a unique_ptr to a concrete object
             ///        constructed using the underlying implementation
-            static std::unique_ptr<SharedMemoryUser> make_unique(const std::string &shm_key, unsigned int timeout);
+            static std::unique_ptr<SharedMemoryUser> make_unique(const std::string &shm_key, double timeout);
             /// @brief Returns a shared_ptr to a concrete object
             ///        constructed using the underlying implementation
-            static std::shared_ptr<SharedMemoryUser> make_shared(const std::string &shm_key, unsigned int timeout);
+            static std::shared_ptr<SharedMemoryUser> make_shared(const std::string &shm_key, double timeout);
     };
 }
